backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
web3==6.15.1
eth-account==0.11.0
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.2.0
boto3==1.34.34
ipfshttpclient==0.8.0a2
django-filter==23.5
drf-spectacular==0.27.1
